+++ packages/orq-rc/RELEASES.md
@@ -9519,3 +9519,13 @@ Based on:
 - [typescript v4.10.0-rc.49] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.49] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.49 - packages/orq-rc
+
+## 2026-05-28 08:20:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.768.2 (2.889.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.50] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.50] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.50 - packages/orq-rc